refactor: extract executionErrorStore from executionStore
Separate error-related state and logic from executionStore into a
dedicated executionErrorStore for better modularity.
Changes:
- Create executionErrorStore with all error state (lastNodeErrors,
  lastExecutionError, lastPromptError), computed properties
  (hasAnyError, totalErrorCount, activeGraphErrorNodeIds), and
  UI state (isErrorOverlayOpen, showErrorOverlay, dismissErrorOverlay)
- Move executionIdToNodeLocatorId to graphTraversalUtil, reusing
  existing traverseSubgraphPath and accepting rootGraph as parameter
- Update 12 consumer files to import from executionErrorStore
- Retain deprecated getters in ComfyApp for backward compatibility
